refactor(airtable): Move the tools and make request code to tools folder
