+++ mcp_servers/airtable/tools/__init__.py
@@ -1,0 +1,9 @@
+# Airtable MCP Server Tools
+# This package contains all the tool implementations organized by object type
+
+from .bases import get_bases_info
+
+__all__ = [
+    # Bases
+    "get_bases_info",
+]
